refactor(app): update project architecture to FSD
