+++ src/app/providers/ThemeProvider/lib/useTheme.tsx
@@ -1,5 +1,9 @@
 import { useContext } from 'react'
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from './themeContext'
+import {
+	Theme,
+	ThemeContext,
+	LOCAL_STORAGE_THEME_KEY,
+} from '../lib/ThemeContext'
 
 interface UseThemeResult {
 	toggleTheme: () => void
