+++ src/app/styles/themes/dark.scss
