--- src/components/pages/AboutPage/AboutPage.lazy.tsx
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export const AboutPageLazy = React.lazy(() => import('./AboutPage'))
