--- src/components/pages/AboutPage/AboutPage.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-function AboutPage() {
-	return (
-		<div>
-			<h1>About Page</h1>
-		</div>
-	)
-}
-export default AboutPage
