--- src/components/pages/MainPage/MainPage.lazy.tsx
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export const MainPagLazy = React.lazy(() => import('./MainPage'))
