--- src/components/pages/MainPage/MainPage.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function MainPage() {
-	return (
-		<div>
-			<h1>Main Page</h1>
-		</div>
-	)
-}
