+++ src/pages/AboutPage/index.ts
@@ -1,0 +1,3 @@
+import { AboutPageLazy } from './ui/AboutPage.lazy'
+
+export { AboutPageLazy as AboutPage }
