+++ src/pages/AboutPage/ui/AboutPage.lazy.tsx
@@ -1,0 +1,3 @@
+import { lazy } from 'react'
+
+export const AboutPageLazy = lazy(() => import('./AboutPage'))
