+++ src/pages/AboutPage/ui/AboutPage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const AboutPage = () => {
+    return (
+        <div>
+            ABOUT PAGE
+        </div>
+    );
+};
+
+export default AboutPage;
