+++ src/pages/MainPage/index.ts
@@ -1,0 +1,3 @@
+import { MainPageLazy } from './ui/MainPage.lazy'
+
+export { MainPageLazy as MainPage }
