+++ src/pages/MainPage/ui/MainPage.lazy.tsx
@@ -1,0 +1,3 @@
+import { lazy } from 'react'
+
+export const MainPageLazy = lazy(() => import('./MainPage'))
