+++ src/pages/MainPage/ui/MainPage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const MainPage = () => {
+    return (
+        <div>
+            MainPage
+        </div>
+    );
+};
+
+export default MainPage;
