+++ tsconfig.json
@@ -14,6 +14,10 @@
 		"moduleResolution": "node",
 		// esModuleInterop позволяет работать с пакетами, которые используют
 		// common js как с обычными пакета с помощью import (require() module.export) = common js
+		"baseUrl": ".",
+		"paths": {
+			"*": ["./src/*"]
+		},
 		// import Moment from 'moment';
 		// без флага esModuleInterop результат undefined
 		// console.log(Moment);
